refactor(iso): Organize data structures into dedicated modules
Moved several ISO-related data structures and constants from `src/iso/builder.rs` to new, more specific modules.

The following types and constants were relocated:
- `IsoFile`, `IsoDirectory`, `IsoFsNode` to `src/iso/fs_node.rs`
- `BootInfo`, `BiosBootInfo`, `UefiBootInfo` to `src/iso/boot_info.rs`
- `IsoImage`, `IsoImageFile` to `src/iso/iso_image.rs`
- `ESP_START_LBA` to `src/iso/constants.rs`

This refactoring improves modularity and code organization by separating concerns. `builder.rs` now focuses solely on the ISO image construction logic, while the new modules define their respective data structures, making them more discoverable and reusable. The public API (`src/lib.rs`) has been updated to re-export these types for external use.
