+++ src/iso/constants.rs
@@ -1,0 +1,2 @@
+/// The starting LBA for the EFI System Partition in isohybrid images.
+pub const ESP_START_LBA: u32 = 34;
