+++ src/iso/mod.rs
@@ -1,8 +1,10 @@
 pub mod boot_catalog;
+pub mod boot_info;
 pub mod builder;
+pub mod constants;
 pub mod dir_record;
+pub mod fs_node;
 pub mod gpt;
+pub mod iso_image;
 pub mod mbr;
 pub mod volume_descriptor;
-
-pub const ESP_START_LBA: u32 = 34;
